requirements.txt
openai
langchain
pdfplumber
pytesseract
pillow
python-docx
tqdm
pylatexenc
faiss-cpu
llama-index
easyocr
langchain_community
groq